requirements.txt
python-dotenv==1.1.0
requests==2.32.4
beautifulsoup4==4.13.4
groq==0.28.0
streamlit==1.45.1
faiss-cpu==1.11.0
sentence-transformers==4.1.0
numpy==2.2.6
pickle-mixin==1.0.2 